move bulk out of lib.rs
